requirements.txt
asgiref==3.11.1
Django==4.2.28
pillow==11.3.0
pillow_heif==1.1.1
sqlparse==0.5.5
typing_extensions==4.15.0